UnitTest changes

+++ Frends.FTP.UploadFiles/Frends.FTP.UploadFiles.Tests/UploadFilesTests.cs
@@ -96,35 +96,19 @@ public class UploadFilesTests
     public void CleanUp()
     {
         if (Directory.Exists(_dir))
+            Directory.Delete(_dir, true);
+        
+        // Permission issues when testing in CI (linux).
+        if (OperatingSystem.IsWindows())
         {
-            try
-            {
-                Directory.Delete(_dir, true);
-                Console.WriteLine("Directory deleted successfully.");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"An error occurred while deleting the directory: {ex.Message}");
-            }
-        }
+            var dirs = new List<string>() { "TestFiles", "DestinationAction_Append_NoRenameInTransfer", "DestinationAction_Append_WithRenameInTransfer", "MacrosWorkInDestinationFileName", "MacrosWorkInSourceDirectory", "SourceOperation_Delete", "SourceOperation_Move", "SourceOperation_Nothing", "SourceOperation_Rename", "SourceOperation_Delete" };
 
-        var dirs = new List<string>() { "TestFiles", "DestinationAction_Append_NoRenameInTransfer", "DestinationAction_Append_WithRenameInTransfer", "MacrosWorkInDestinationFileName", "MacrosWorkInSourceDirectory", "SourceOperation_Delete", "SourceOperation_Move", "SourceOperation_Nothing", "SourceOperation_Rename", "SourceOperation_Delete" };
-
-        foreach (var dir in dirs)
-        {
-            var targetDir = Path.Combine(Directory.GetCurrentDirectory(), "..", "..", "..", "DockerVolumes", "data", dir);
-            var fullPath = Path.GetFullPath(targetDir);
-            if (Directory.Exists(fullPath))
+            foreach (var dir in dirs)
             {
-                try
-                {
-                    Directory.Delete(fullPath, true);
-                    Console.WriteLine($"Directory '{fullPath}' deleted successfully.");
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"An error occurred while deleting directory '{fullPath}': {ex.Message}");
-                }
+                var targetDir = Path.Combine(Directory.GetCurrentDirectory(), "..", "..", "..", "DockerVolumes", "data", dir);
+                var fullPath = Path.GetFullPath(targetDir);
+                if (Directory.Exists(fullPath) && OperatingSystem.IsWindows())
+                        Directory.Delete(fullPath, true);
             }
         }
     }
